[Refactor] rename files with `classic` prefix
