+++ src/workbench/services/fileTree/fileItem.ts
