+++ src/workbench/services/fileTree/fileItemRenderer.ts
@@ -2,7 +2,7 @@ import { ITreeNode } from "src/base/browser/secondary/tree/tree";
 import { ITreeListRenderer } from "src/base/browser/secondary/tree/treeListRenderer";
 import { IListViewMetadata, RendererType } from "src/base/browser/secondary/listView/listRenderer";
 import { FuzzyScore } from "src/base/common/fuzzy";
-import { FileItem } from "src/workbench/services/classicTree/classicItem";
+import { FileItem } from "src/workbench/services/fileTree/fileItem";
 import { Icons } from "src/base/browser/icon/icons";
 import { getIconClass } from "src/base/browser/icon/iconRegistry";
 import { IListItemProvider } from "src/base/browser/secondary/listView/listItemProvider";
